LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-row-axis-self-baseline-synthesized-004-expected.txt

Horizontal grid and verticalLR item with fixed width
É É ÉÉ
 Horizontal grid and verticalRL item with fixed width
É É ÉÉ
 VerticalLR grid and item with fixed height
É É ÉÉ
 VerticalRL grid and item with fixed height
É É ÉÉ

PASS .inline-grid 1
FAIL .inline-grid 2 assert_equals:
<div class="inline-grid justifyItemsBaseline" data-expected-width="150" data-expected-height="300">
  <div class="firstRowFirstColumn fixedWidth" data-offset-x="0" data-offset-y="0" data-expected-width="125" data-expected-height="100"></div>
  <div class="secondRowFirstColumn bigFont paddingRight verticalRL" data-offset-x="5" data-offset-y="100" data-expected-width="120" data-expected-height="200">É É ÉÉ</div>
  <div class="autoRowSpanning2AutoColumn width25"></div>
</div>
width expected 150 but got 210
PASS .inline-grid 3
PASS .inline-grid 4
